Clean up in-app overlay when app resumes after handling an open URL button action
When a button action for in-app messages includes opening a URL, the current activity is paused
which makes the ForegroundStateWatcher return null in presentMessagesToClient.

This in turn has the effect that the last message in the queue is not cleaning up the view when
the user navigates back to the app, and it appears the app is 'frozen' because the in-app overlay
is still on top of the screen.

The fix checks if the queue is empty in the public presentMessages routine, which is called when
the app resumes any activity, and then cleans up the view if there is nothing more to present.