Revert specification of FCM library dependency as a range (#63)
Range version checking does not work correctly when the Google Services gradle plugin is applied to consumer
app projects. This seems to be a bug/limitation in the Google Services gradle plugin logic.

This results in an error that can only be resolved by disabling version checking, which defeats the purpose
of supporting ranges of compatible versions of the FCM library.

As the Firebase libraries are moving towards Bill of Materials (BoM) specifications for dependency versioning,
revert our previous change in favour of supporting BoM more explicitly in an upcoming release.